requirements.txt
Click==7.0
Flask==1.0.2
Flask-Login==0.4.1
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
Flask-Script==2.0.6
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
SQLAlchemy==1.2.14
WTForms==2.2.1
Werkzeug==0.14.1
alembic==1.0.3
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.1
gunicorn==19.9.0
idna==2.7
itsdangerous==1.1.0
oauthlib==2.1.0
psycopg2==2.7.6.1
pyOpenSSL==18.0.0
pycparser==2.19
python-dateutil==2.7.5
python-dotenv==0.9.1
python-editor==1.0.3
requests==2.20.1
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.24.1
